test(hosted): split target adapter coverage
